fix(ui): implement state-based deep linking for OIDC authentication
Fixes deep linking issues with Entra ID authentication by leveraging
state-based navigation instead of relying on IDP redirect URIs.

- Updated @apicurio/common-ui-components to 2.1.1
- Refactored App.tsx to separate authentication concerns
- Created MainPageWithAuth component with onRedirect callback
- Removed unused CSS styles from App.css
- Updated build-docker.sh to include build steps

This change enables proper deep linking with OIDC providers that
restrict wildcard redirect URIs, particularly Microsoft Entra ID.

Fixes #6364
