+++ ui/build-docker.sh
@@ -5,11 +5,10 @@ then
   DOCKER_CMD="docker"
 fi
 
-if [ ! -d "dist" ]
-then
-  echo "Missing 'dist' directory.  Build first!"
-  exit 1
-fi
+npm install
+npm run clean
+npm run build
+npm run package
 
 $DOCKER_CMD build -t="apicurio/apicurio-registry-ui:latest-snapshot" --rm .
