+++ ui/ui-app/package.json
@@ -38,7 +38,7 @@
   },
   "dependencies": {
     "@apicurio/apicurio-registry-sdk": "3.1.7-Dev",
-    "@apicurio/common-ui-components": "2.1.0",
+    "@apicurio/common-ui-components": "2.1.1",
     "@apicurio/data-models": "1.1.33",
     "@microsoft/kiota-abstractions": "1.0.0-preview.99",
     "@microsoft/kiota-http-fetchlibrary": "1.0.0-preview.99",
