+++ ui/ui-app/src/app/App.css
@@ -2,17 +2,6 @@
   --myApp-global--spacer--md: 30px;
 }
 
-.app-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-top: var(--myApp-global--spacer--md);
-}
-
-.notification-container {
-  flex: 0 1 auto;
-}
-
 .pf-c-page {
   width: 100%;
   height: 100vh;
